requirements.txt
fastapi==0.110.0
pydantic==2.9.2
google-generativeai==0.7.2
python-multipart==0.0.9
jinja2==3.1.4
uvicorn[standard]==0.24.0
websockets==12.0
python-dotenv==1.0.1